# openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Kartówka API",
    "description": "The Kartówka REST-like API.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/questions": {
      "post": {
        "tags": [
          "Questions"
        ],
        "summary": "Creates a new question",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQuestionDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQuestionDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQuestionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "New question object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Question"
                }
              }
            }
          },
          "400": {
            "description": "Question validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Question"
                }
              }
            }
          }
        }
      }
    },
    "/api/questions/{questionId}": {
      "get": {
        "tags": [
          "Questions"
        ],
        "summary": "Get question by ID",
        "parameters": [
          {
            "name": "questionId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Question model",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Question"
                }
              }
            }
          },
          "404": {
            "description": "Not found error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Questions"
        ],
        "summary": "Update a question",
        "parameters": [
          {
            "name": "questionId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateQuestionDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateQuestionDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateQuestionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated question object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Question"
                }
              }
            }
          },
          "400": {
            "description": "Question validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Question not found error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Questions"
        ],
        "summary": "Removes a question",
        "description": "This endpoint is idempotent.",
        "parameters": [
          {
            "name": "questionId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/questions-categories": {
      "post": {
        "tags": [
          "QuestionsCategories"
        ],
        "summary": "Create a new question",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQuestionsCategoryDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQuestionsCategoryDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQuestionsCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "New questions category object.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionsCategory"
                }
              }
            }
          },
          "400": {
            "description": "Validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/questions-categories/{categoryId}": {
      "patch": {
        "tags": [
          "QuestionsCategories"
        ],
        "summary": "Create a new questions category",
        "parameters": [
          {
            "name": "categoryId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateQuestionsCategoryDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateQuestionsCategoryDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateQuestionsCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated questions category object.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionsCategory"
                }
              }
            }
          },
          "400": {
            "description": "Validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "QuestionsCategories"
        ],
        "summary": "Remove a category",
        "parameters": [
          {
            "name": "categoryId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No content success response."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Asset": {
        "required": [
          "blobUrl",
          "displayName",
          "systemName"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "displayName": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "systemName": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "size": {
            "maximum": 2147483647,
            "minimum": 0,
            "type": "integer",
            "format": "int64"
          },
          "assetType": {
            "$ref": "#/components/schemas/AssetType"
          },
          "blobUrl": {
            "maxLength": 400,
            "minLength": 0,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "AssetType": {
        "enum": [
          "image",
          "music"
        ],
        "type": "string"
      },
...
      "CreateQuestionDto": {
        "required": [
          "answer"
        ],
        "type": "object",
        "properties": {
          "packId": {
            "type": "integer",
            "format": "int64"
          },
          "questionsCategoryId": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "questionText": {
            "maxLength": 400,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "assetId": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "score": {
            "maximum": 2147483647,
            "minimum": 0,
            "type": "integer",
            "format": "int32"
          },
          "answer": {
            "maxLength": 400,
            "minLength": 0,
            "type": "string"
          },
          "contentType": {
            "$ref": "#/components/schemas/QuestionContentType"
          },
          "questionType": {
            "$ref": "#/components/schemas/QuestionType"
          }
        },
        "additionalProperties": false
      },
      "CreateQuestionsCategoryDto": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "packId": {
            "type": "integer",
            "format": "int64"
          },
          "roundId": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "order": {
            "maximum": 2147483647,
            "minimum": 0,
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationResult"
            },
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
      "Question": {
        "required": [
          "answer"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "questionsCategoryId": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "questionText": {
            "maxLength": 400,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "score": {
            "maximum": 2147483647,
            "minimum": 0,
            "type": "integer",
            "format": "int32"
          },
          "answer": {
            "maxLength": 400,
            "minLength": 0,
            "type": "string"
          },
          "contentType": {
            "$ref": "#/components/schemas/QuestionContentType"
          },
          "questionType": {
            "$ref": "#/components/schemas/QuestionType"
          },
          "asset": {
            "$ref": "#/components/schemas/Asset"
          }
        },
        "additionalProperties": false
      },
      "QuestionContentType": {
        "enum": [
          "text",
          "image",
          "music"
        ],
        "type": "string"
      },
      "QuestionType": {
        "enum": [
          "default",
          "pass",
          "auction",
          "sponsored"
        ],
        "type": "string"
      },
      "QuestionsCategory": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "roundId": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "order": {
            "maximum": 2147483647,
            "minimum": 0,
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "UpdateQuestionDto": {
        "type": "object",
        "properties": {
          "questionsCategoryId": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "questionText": {
            "maxLength": 400,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "assetId": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "score": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "answer": {
            "maxLength": 400,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "contentType": {
            "$ref": "#/components/schemas/QuestionContentType"
          },
          "questionType": {
            "$ref": "#/components/schemas/QuestionType"
          }
        },
        "additionalProperties": false
      },
      "UpdateQuestionsCategoryDto": {
        "type": "object",
        "properties": {
          "roundId": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "order": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "name": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ValidationResult": {
        "type": "object",
        "properties": {
          "memberNames": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true,
            "readOnly": true
          },
          "errorMessage": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Put JWT token here",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}